+++ bgplot/entities/__init__.py
@@ -3,3 +3,12 @@ from .vector import Vector
 from .axes import Axes
 from .line import Line
 from .plane import Plane
+from .oriented_point import OrientedPoint
+
+__all__ = ['Point',
+           'Vector',
+           'Axes',
+           'Line',
+           'Plane',
+           'OrientedPoint',
+           ]